Removed old files, organized files into folders, split Ride up
