+++ lib/models/Rider.dart
@@ -1,4 +1,4 @@
-import 'package:carriage/app_config.dart';
+import '../utils/app_config.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'dart:convert';
