+++ lib/pages/ride-flow/PickUpPage.dart
@@ -1,14 +1,15 @@
-import 'package:carriage/MeasureRect.dart';
-import 'package:carriage/Ride.dart';
+import '../../utils/MeasureRect.dart';
 import 'package:carriage/widgets/Buttons.dart';
 import 'package:carriage/widgets/Dialogs.dart';
 import 'package:carriage/widgets/RideStops.dart';
 import 'package:flutter/material.dart';
 import 'package:loading_overlay/loading_overlay.dart';
 import 'package:provider/provider.dart';
-import '../CarriageTheme.dart';
+import '../../utils/CarriageTheme.dart';
 import '../Home.dart';
-import '../RidesProvider.dart';
+import '../../providers/RidesProvider.dart';
+import '../../models/Ride.dart';
+import '../../models/Rider.dart';
 
 class PickUpPage extends StatefulWidget {
   final OnWidgetRectChange onContinueRectChange;
