+++ lib/providers/RidesProvider.dart
@@ -4,8 +4,8 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 import 'AuthProvider.dart';
-import 'Ride.dart';
-import 'app_config.dart';
+import '../models/Ride.dart';
+import '../utils/app_config.dart';
 import 'package:http/http.dart' as http;
 
 class RidesProvider with ChangeNotifier {
