+++ lib/widgets/RideDestPickupCard.dart
@@ -1,4 +1,4 @@
-import 'package:carriage/CarriageTheme.dart';
+import '../utils/CarriageTheme.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
